Working, but sloppy.

+++ srcs/main.c
@@ -6,30 +6,14 @@
 /*   By: jbrinksm <jbrinksm@student.codam.nl>         +#+                     */
 /*                                                   +#+                      */
 /*   Created: 2019/04/10 20:29:49 by jbrinksm       #+#    #+#                */
-/*   Updated: 2019/10/21 14:24:59 by rkuijper      ########   odam.nl         */
+/*   Updated: 2019/10/24 10:30:09 by rkuijper      ########   odam.nl         */
 /*                                                                            */
 /* ************************************************************************** */
 
 #include "vsh.h"
 #include <signal.h>
 
-void			signal_handle(int sig)
-{
-	t_job *job;
-
-	if (sig == SIGTSTP)
-	{
-		job = g_data->jobs->joblist;
-		while (job != NULL)
-		{
-			kill(job->pgid, SIGTSTP);
-			job = job->next;
-		}
-		ft_putchar('\n');
-	}
-}
-
-void	init_g_state(void)
+static void	init_shell(void)
 {
 	g_state = (t_state*)ft_memalloc(sizeof(t_state));
 	if (g_state == NULL)
@@ -37,28 +21,32 @@ void	init_g_state(void)
 		ft_eprintf(E_ALLOC_STR);
 		exit(EXIT_FAILURE);
 	}
-	g_state->pid = getpid();
 	g_state->shell_type = SHELL_INTERACT;
-	while (tcgetpgrp(STDIN_FILENO) != g_state->pid)
-		kill(-g_state->pid, SIGTTIN);
-	signal(SIGINT, SIG_IGN);
-	signal(SIGQUIT, SIG_IGN);
-	signal(SIGTSTP, SIG_IGN);
-	signal(SIGTTIN, SIG_IGN);
-	signal(SIGTTOU, SIG_IGN);
-	g_state->pid = getpid();
-	setpgid(g_state->pid, g_state->pid);
-	tcsetpgrp(STDIN_FILENO, g_state->pid);
+	if (isatty(STDIN_FILENO))
+	{
+		g_state->pid = getpgrp();
+		while (tcgetpgrp(STDIN_FILENO) != g_state->pid)
+		{
+			kill(-g_state->pid, SIGTTIN);
+			g_state->pid = getpgrp();
+		}
+		signal(SIGINT, SIG_IGN);
+		signal(SIGQUIT, SIG_IGN);
+		signal(SIGTSTP, SIG_IGN);
+		signal(SIGTTIN, SIG_IGN);
+		signal(SIGTTOU, SIG_IGN);
+		signal(SIGCHLD, signal_handle_child_death);
+		g_state->pid = getpid();
+		setpgid(g_state->pid, g_state->pid);
+		tcsetpgrp(STDIN_FILENO, g_state->pid);
+	}
 }
 
-int		main(int argc, char **argv)
+int			main(int argc, char **argv)
 {
 	t_vshdata	*data;
 
-	signal(SIGTSTP, signal_handle);
-	signal(SIGCHLD, signal_handle_child_death);
-	signal(SIGINT, SIG_IGN);
-	init_g_state();
+	init_shell();
 	data = shell_init_vshdata();
 	if (data == NULL)
 		return (EXIT_FAILURE);
